+++ WorkoutService/Controllers/WorkoutController.cs
@@ -25,35 +25,35 @@ namespace WorkoutService.Controllers
 
         [Route("users/{userId}")]
         [HttpGet]
-        public async Task<IActionResult> GetAll(int userId)
+        public async Task<IActionResult> GetAll(long userId)
         {
             return Ok(new Response<IEnumerable<WorkoutDTO>>(await _workoutService.GetAllWorkouts(userId)));
         }
 
         [Route("users/{userId}/search")]
         [HttpGet]
-        public async Task<IActionResult> SearchWorkouts(int userId, string type, int page, int size)
+        public async Task<IActionResult> SearchWorkouts(long userId, string type, int page, int size)
         {
             return Ok(new Response<IEnumerable<WorkoutDTO>>(await _workoutService.SearchWorkouts(userId, type, page, size)));
         }
 
         [Route("users/{userId}/report")]
         [HttpGet]
-        public async Task<IActionResult> SearchForReport(int userId, string fromDate, string toDate)
+        public async Task<IActionResult> SearchForReport(long userId, string fromDate, string toDate)
         {
             return Ok(new Response<IEnumerable<WorkoutDTO>>(await _workoutService.SearchForReport(userId, fromDate, toDate)));
         }
 
         [Route("users/{userId}/workouts/{id}")]
         [HttpGet]
-        public async Task<IActionResult> GetWorkoutById(int userId, int id)
+        public async Task<IActionResult> GetWorkoutById(long userId, int id)
         {
             return Ok(new Response<WorkoutDTO>(await _workoutService.GetWorkoutById(userId, id)));
         }
 
         [Route("users/{userId}/workouts/{id}")]
         [HttpDelete]
-        public void RemoveWorkout(int userId, int id)
+        public void RemoveWorkout(long userId, long id)
         {
             _workoutService.RemoveWorkout(userId, id);
         }
