+++ WorkoutService/Controllers/WorkoutTypeController.cs
@@ -24,7 +24,7 @@ namespace WorkoutService.Controllers
 
         [Route("{id}")]
         [HttpGet]
-        public async Task<IActionResult> GeById(int id)
+        public async Task<IActionResult> GeById(long id)
         {
             return Ok(new Response<WorkoutTypeDTO>(await _workoutTypeService.GetWorkoutTypeById(id)));
         }
