+++ WorkoutService/Services/IWorkoutService.cs
@@ -4,16 +4,16 @@ namespace WorkoutService.Services
 {
     public interface IWorkoutService
     {
-        Task<IEnumerable<WorkoutDTO>> GetAllWorkouts(int userId);
+        Task<IEnumerable<WorkoutDTO>> GetAllWorkouts(long userId);
 
-        Task<WorkoutDTO> GetWorkoutById(int userId, int id);
+        Task<WorkoutDTO> GetWorkoutById(long userId, long id);
 
         Task<WorkoutDTO> LogWorkout(WorkoutDTO workoutDTO);
 
-        Task<IEnumerable<WorkoutDTO>> SearchWorkouts(int userId, string workoutType, int page, int size);
+        Task<IEnumerable<WorkoutDTO>> SearchWorkouts(long userId, string workoutType, int page, int size);
 
-        Task<IEnumerable<WorkoutDTO>> SearchForReport(int userId, string fromDate, string toDate);
+        Task<IEnumerable<WorkoutDTO>> SearchForReport(long userId, string fromDate, string toDate);
 
-        void RemoveWorkout(int userId, int id);
+        void RemoveWorkout(long userId, long id);
     }
 }
